refactor: Move where we export `PressProps`
