+++ mobile/src/components/Base/Pressable.tsx
@@ -6,3 +6,14 @@ import { Pressable as RNPresasble } from "react-native";
 
 export type PressableProps = RNPressableProps;
 export const Pressable = RNPresasble;
+
+/** The most "used" action props used on `<Pressable />`. */
+export type PressProps = Pick<
+  PressableProps,
+  | "disabled"
+  | "delayLongPress"
+  | "onLongPress"
+  | "onPress"
+  | "onPressIn"
+  | "onPressOut"
+>;
